requirements.txt
# Flattened dev requirements for Vercel (auto-generated from requirements/common.txt + requirements/dev.txt)
Babel==2.18.0
django-admin-dracula==0.0.3
django-contact-form==5.2.0
django-countries==8.2.0
django-hosts==7.0.0
django-money==3.6.0
django-push @ git+https://github.com/brutasse/django-push.git@22fda99641cfbd2f3075a723d92652a8e38220a5
django-read-only==1.21.0
django-recaptcha==4.1.0
django-registration-redux==2.13
Django==6.0.2
docutils==0.21.2
feedparser==6.0.12
Jinja2==3.1.6
libsass==0.23.0
Markdown==3.10.2
Pillow==12.1.1
Pygments==2.19.2
pykismet3==0.1.1
pymdown-extensions==10.21
python-dateutil==2.9.0.post0
requests==2.32.5
sorl-thumbnail==12.11.0
Sphinx==8.1.3
stripe==13.2.0
time-machine==3.2.0
black==26.1.0
django-debug-toolbar==6.2.0
prek==0.3.3
psycopg[binary]==3.3.2
watchdog==6.0.0
whitenoise==6.9.0
